Merged PR 2953: (13746)[Backend]Feat: adding a new feature that analyzes the candidate resume...
This feature takes a PDF file, request a Text Cleaning Service and make a analysis on candidate resume. To use this service, we add a route to the PDF Controller and two files to the models, one that vectorizes the text and the other with an AI-trained model.

Related work items: #13746