Merged PR 2995: (13840)[Backend]Fix: improve the level of reliability of the return
We improve the reliability level with a new AI trained model, and to use him, we add a mapping called classMapping. To guarantee a quality return, we define a limit of 30% position probability, positions with less than 30% will not be include on him. On this version we also update the requirements and delete the CleaningService.

Related work items: #13840